requirements.txt
adium-theme-ubuntu==0.3.4
backports.ssl-match-hostname==3.4.0.2
cached-property==1.3.0
certifi==2017.7.27.1
chardet==3.0.4
cryptography==1.2.3
Django==1.11.5
django-filter==1.0.4
djangorestframework==3.6.4
docker-compose==1.8.0
docker-py==1.9.0
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.2
funcsigs==0.4
functools32==3.2.3.post2
gyp==0.1
httpie==0.9.9
idna==2.6
ipaddress==1.0.16
jsonschema==2.5.1
Markdown==2.6.9
mock==1.3.0
ndg-httpsclient==0.4.0
pbr==1.8.0
pyasn1==0.1.9
Pygments==2.2.0
pyOpenSSL==0.15.1
pytz==2017.2
PyYAML==3.11
requests==2.18.4
six==1.10.0
texttable==0.8.1
unity-lens-photos==1.0
urllib3==1.22
websocket-client==0.18.0
whitenoise==3.3.1
